Add 'HIR' for types
I.e. another intermediate representation between `TypeRef` and `Ty`.

This representation
 - has all names resolved
 - represents bounds for `impl Trait` out of line already
 - but can still only represent actual nameable Rust types, not types that only
   occur after inference like closure types, type variables or bound variables.

I.e. the intention is that this then gets lowered to a `Ty` that only exists
within a specific inference context; this also prepares us for replacing our
`Ty` by Chalk's version.
(TODO maybe it needs to be able to represent unnameable, but
relevant-outside-of-inference types like closure types then)